+++ src/main/java/com/github/camilormz/finalreality/model/character/Enemy.java
@@ -13,6 +13,7 @@ import org.jetbrains.annotations.NotNull;
 public class Enemy extends AbstractCharacter {
 
   private final int weight;
+  private final int damage;
 
   /**
    * Creates a new enemy.
@@ -21,15 +22,21 @@ public class Enemy extends AbstractCharacter {
    *     the enemy's name
    * @param weight
    *     the enemy's weight, variable relevant for time enqueuing
+   * @param healthPoints
+   *     the enemy's health points
+   * @param damage
+   *     the damage the enemy is able to do in a single attack
    * @param turnsQueue
    *     the queue with the characters waiting for their turn
    */
   public Enemy(@NotNull final String name,
                final int weight,
                int healthPoints,
+               final int damage,
                @NotNull final BlockingQueue<ICharacter> turnsQueue) {
     super(turnsQueue, name, healthPoints, CharacterDomain.ENEMY);
     this.weight = weight;
+    this.damage = damage;
   }
 
   /**
@@ -45,6 +52,11 @@ public class Enemy extends AbstractCharacter {
   }
 
   @Override
+  public int getDamagePoints() {
+    return this.damage;
+  }
+
+  @Override
   public boolean equals(final Object o) {
     if (this == o) {
       return true;
@@ -54,7 +66,8 @@ public class Enemy extends AbstractCharacter {
     }
     final Enemy enemy = (Enemy) o;
     return this.getName().equals(enemy.getName()) &&
-           this.getWeight() == enemy.getWeight();
+           this.getWeight() == enemy.getWeight() &&
+           this.getDamagePoints() == enemy.getDamagePoints();
   }
 
   @Override