+++ src/main/java/com/github/camilormz/finalreality/model/character/player/characterclass/BlackMage.java
@@ -21,6 +21,8 @@ public class BlackMage extends AbstractPlayerCharacter implements IMagicalCharac
      *
      * @param name
      *     the black mage's name
+     * @para healthPoints
+     *     the black mage's health points
      * @param turnsQueue
      *     the queue with the characters waiting for their turn
      */